Fix macOS service installation launch template
The current version of cloudflared (2021.3.2 at time of commit)
requires additional arguments to start a tunnel. Running 'cloudflared'
on its own does not run a tunnel with the default configuration.

This is behavior that has changed from previous versions of cloudflared,
where no arguments were required for starting/running a tunnel with the
default configuration behavior.