A Helm chart for a cloudflare tunnel.
This is in reference to github issue #569. I saw the example yaml in
https://github.com/cloudflare/argo-tunnel-examples/tree/master/named-tunnel-k8s
and thought it would be much more useful as a helm chart. And I had to
make a helm chart anyway so I could deploy it with Terraform.  Here's
the result, and figured I'd contribute it back.