Use ChangeLog date instead of build date
to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.
This date call only works with GNU date and BSD date.

This patch was done while working on reproducible builds for openSUSE, sponsored by the NLnet NGI0 fund.