fix: expand home directory for credentials file
The documentation does not specify that the `credentials-file` field in
`config.yaml` needs to be an absolute path. If you add a relative path
with a `~`, it throw as error that the file is not found. However, this
iis a confusing error for the user as the file at the path exists,
it is just that `os.Stat` failed because it could not expand the `~`.
This commit fixes the above issue by running a `homedir.Expand` on the
`credentials-file` path in the `credentialFinder` function.